# api_gateway/api/docs/swagger.yaml
basePath: /
definitions:
...
  models.CreateTransaction:
    properties:
      account_id:
        description: Account ID as a string
        type: string
      amount:
        description: Amount as a float32
        type: number
      category_id:
        description: Category ID as a string
        type: string
      date:
        description: Date as a time.Time
        type: string
      description:
        description: Description as a string
        type: string
      type:
        description: Type (spending/income) as a string
        type: string
      user_id:
        description: User ID as a string
        type: string
    type: object
...
  models.GetUserMoneyResponse:
    properties:
      category_id:
        description: Category ID as a string
        type: string
      time:
        description: Time as a time.Time
        type: string
      total_amount:
        description: Total amount as a float32
        type: number
    type: object
  models.GetUserMoneysResponse:
    properties:
      moneys:
        description: List of user money responses
        items:
          $ref: '#/definitions/models.GetUserMoneyResponse'
        type: array
    type: object
...
  models.Response:
    properties:
      data: {}
      description:
        type: string
      statusCode:
        type: integer
    type: object
  models.Transaction:
    properties:
      account_id:
        description: Account ID as a string
        type: string
      amount:
        description: Amount as a float32
        type: number
      category_id:
        description: Category ID as a string
        type: string
      created_at:
        description: Created timestamp as a time.Time
        type: string
      date:
        description: Date as a time.Time
        type: string
      description:
        description: Description as a string
        type: string
      id:
        description: ObjectID for MongoDB
        type: string
      type:
        description: Type (spending/income) as a string
        type: string
      updated_at:
        description: Updated timestamp as a time.Time
        type: string
      user_id:
        description: User ID as a string
        type: string
    type: object
  models.TransactionsResponse:
    properties:
      count:
        description: Count of transactions
        type: integer
      transactions:
        description: List of transactions
        items:
          $ref: '#/definitions/models.Transaction'
        type: array
    type: object
...
  models.UpdateTransaction:
    properties:
      account_id:
        description: Account ID as a string
        type: string
      amount:
        description: Amount as a float32
        type: number
      category_id:
        description: Category ID as a string
        type: string
      date:
        description: Date as a time.Time
        type: string
      description:
        description: Description as a string
        type: string
      id:
        description: ObjectID for MongoDB
        type: string
      type:
        description: Type (spending/income) as a string
        type: string
      user_id:
        description: User ID as a string
        type: string
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Personal Finance Tracker API Documentation
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Personal Finance Tracker API
  version: "1.0"
paths:
...
  /budget_service/v5/transaction:
    post:
      consumes:
      - application/json
      description: Create a new transaction
      parameters:
      - description: transaction
        in: body
        name: transaction
        required: true
        schema:
          $ref: '#/definitions/models.CreateTransaction'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Transaction'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - BearerAuth: []
      summary: Create a new transaction
      tags:
      - transaction
  /budget_service/v5/transaction/{id}:
    delete:
      consumes:
      - application/json
      description: Delete transaction
      parameters:
      - description: transaction_id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: Delete transaction
      tags:
      - transaction
    get:
      consumes:
      - application/json
      description: Get transaction by id
      parameters:
      - description: transaction_id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Transaction'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: Get transaction by id
      tags:
      - transaction
    put:
      consumes:
      - application/json
      description: Update transaction
      parameters:
      - description: transaction_id
        in: path
        name: id
        required: true
        type: string
      - description: transaction
        in: body
        name: transaction
        schema:
          $ref: '#/definitions/models.UpdateTransaction'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Transaction'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: Update transaction
      tags:
      - transaction
  /budget_service/v5/transactions:
    get:
      consumes:
      - application/json
      description: Get all transaction
      parameters:
      - description: page
        in: query
        name: page
        type: string
      - description: limit
        in: query
        name: limit
        type: string
      - description: search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TransactionsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: Get all transaction
      tags:
      - transaction
  /budget_service/v5/transactions/income:
    get:
      consumes:
      - application/json
      description: Get all transaction
      parameters:
      - description: user_id
        in: query
        name: user_id
        type: string
      - description: start_time
        in: query
        name: start_time
        type: string
      - description: end_time
        in: query
        name: end_time
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GetUserMoneysResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: Get all transaction
      tags:
      - transaction
  /budget_service/v5/transactions/spend:
    get:
      consumes:
      - application/json
      description: Get all transaction
      parameters:
      - description: user_id
        in: query
        name: user_id
        type: string
      - description: start_time
        in: query
        name: start_time
        type: string
      - description: end_time
        in: query
        name: end_time
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GetUserMoneysResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: Get all transaction
      tags:
      - transaction
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"